feature:实例维护session，timeout数据结构调整

+++ src/backend/ci/core/websocket/biz-websocket/src/main/kotlin/com/tencent/devops/websocket/cron/ClearTimeoutCron.kt
@@ -82,35 +82,42 @@ class ClearTimeoutCron(
     @Scheduled(cron = "0 */1 * * * ?")
     fun newClearTimeoutCache() {
         val nowTime = System.currentTimeMillis()
-        for (bucket in 0..WebsocketKeys.REDIS_MO){
-            val redisData = redisOperation.get(WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY+bucket)
-            if(redisData != null){
+        for (bucket in 0..WebsocketKeys.REDIS_MO) {
+            val redisData = redisOperation.get(WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY + bucket)
+            if (redisData != null) {
                 val newSessionList = mutableListOf<String>()
                 val sessionList = redisData.split(",")
-                if(sessionList == null || sessionList.isEmpty()){
+                if (sessionList == null || sessionList.isEmpty()) {
+                    logger.info("this bucket is empty,redisKey[${WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY + bucket}]")
                     continue
                 }
                 sessionList.forEach {
-                    logger.info("clearTimeout redisStr[$it]")
-                    val timeout: Long = it.substringAfter("&").toLong()
-                    val userId = it.substringAfter("#").substringBefore("&")
-                    val sessionId = it.substringBefore("#")
-                    logger.info("clearTimeout str[$it] timeout[$timeout] userId[$userId] sessionId[$sessionId]")
-                    if (nowTime > timeout) {
-                        val sessionPage = RedisUtlis.getPageFromSessionPageBySession(redisOperation, sessionId)
-                        RedisUtlis.cleanSessionPageBySessionId(redisOperation, sessionId)
-                        if (sessionPage != null) {
-                            RedisUtlis.cleanPageSessionBySessionId(redisOperation, sessionId, sessionPage)
+                    logger.info("clearTimeout redisStr[$it],redisKey[${WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY + bucket}")
+                    if (it != null && it.isNotEmpty()) {
+                        val timeout: Long = it.substringAfter("&").toLong()
+                        val userId = it.substringAfter("#").substringBefore("&")
+                        val sessionId = it.substringBefore("#")
+                        logger.info("clearTimeout str[$it] timeout[$timeout] userId[$userId] sessionId[$sessionId]")
+                        if (nowTime > timeout) {
+                            val sessionPage = RedisUtlis.getPageFromSessionPageBySession(redisOperation, sessionId)
+                            RedisUtlis.cleanSessionPageBySessionId(redisOperation, sessionId)
+                            if (sessionPage != null) {
+                                RedisUtlis.cleanPageSessionBySessionId(redisOperation, sessionId, sessionPage)
+                            }
+                            RedisUtlis.cleanUserSessionBySessionId(redisOperation, userId, sessionId)
+                            websocketService.removeCacheSession(sessionId)
+                            logger.info("[clearTimeOutSession] sessionId:$sessionId,loadPage:$sessionPage,userId:$userId")
                         }
-                        RedisUtlis.cleanUserSessionBySessionId(redisOperation, userId, sessionId)
-                        websocketService.removeCacheSession(sessionId)
-                        logger.info("[clearTimeOutSession] sessionId:$sessionId,loadPage:$sessionPage,userId:$userId")
+                        newSessionList.add(it)
                     }
-                    newSessionList.add(it)
                 }
-                redisOperation.set(WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY+bucket,objectMapper.writeValueAsString(newSessionList),null, true)
+                redisOperation.set(
+                    WebsocketKeys.HASH_USER_TIMEOUT_REDIS_KEY + bucket,
+                    objectMapper.writeValueAsString(newSessionList),
+                    null,
+                    true
+                )
             }
         }
-
     }
 }
